Refactored the time measurement into a separate header file. Added docs to all implemented functions.
