+++ src/moleculeSimulator/particleGeneration/ParticleGenerator.cpp
@@ -11,7 +11,7 @@ const double ParticleGenerator::brownianMotionAverageVelocity = 0.1;
 int ParticleGenerator::id = 1;
 
 void ParticleGenerator::generateCuboid(ParticleContainer &particles, const std::array<double, 3>& position, unsigned N1,
-                                       unsigned N2, unsigned N3, double h, double mass,const std::array<double, 3>& velocity) {
+                                       unsigned N2, unsigned N3, double h, double mass,const std::array<double, 3>& initVelocity) {
 
     spdlog::info("Generate cuboid with the following parameters:\n"
                  "Position: {}\n"
@@ -21,7 +21,7 @@ void ParticleGenerator::generateCuboid(ParticleContainer &particles, const std::
                  "h: {}\n"
                  "m: {}\n"
                  "Velocity: {}\n",
-                 ArrayUtils::to_string(position), N1, N2, N3, h, mass, ArrayUtils::to_string(velocity)
+                 ArrayUtils::to_string(position), N1, N2, N3, h, mass, ArrayUtils::to_string(initVelocity)
                  );
 
     std::array<double, 3> currentPosition = position;
@@ -30,7 +30,7 @@ void ParticleGenerator::generateCuboid(ParticleContainer &particles, const std::
         for(unsigned n2 = 0; n2 < N2; n2++) {
             for(unsigned n3 = 0; n3 < N3; n3++) {
                 Particle newParticle {currentPosition,
-                    maxwellBoltzmannDistributedVelocity(brownianMotionAverageVelocity,3) + velocity, mass, id};
+                    maxwellBoltzmannDistributedVelocity(brownianMotionAverageVelocity,3) + initVelocity, mass, id};
                 particles.add(newParticle);
                 currentPosition[2] += h;
             }
